Rename package in tests
- The change in package name from `akka.cluster.pi` to `akkapi.cluster`
  in a previous PR was not applied for the test correct. This PR corrects
  that
