+++ exercise_012_clustered_sudoku_solver/src/test/scala/akkapi/cluster/BaseAkkaSpec.scala
@@ -1,4 +1,4 @@
-package akka.cluster.pi
+package akkapi.cluster
 
 import akka.actor.testkit.typed.scaladsl.ActorTestKit
 import org.scalatest.BeforeAndAfterAll
