+++ exercise_012_clustered_sudoku_solver/src/test/scala/akkapi/cluster/ReductionRuleSpec.scala
@@ -1,4 +1,4 @@
-package akka.cluster.pi
+package akkapi.cluster
 
 import org.scalatest.{Matchers, WordSpec}
